chanstate: move kv serialization to stores
Move the generic log update serialization helpers into the KV store
files that use them. Keep the shared byte order with the codec, place
commit-diff log update list serialization with commitment storage, and
give forwarding and revocation logs their own key helpers.
